pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mlops_imdb"
version = "0.0.1"
description = "Film review sentiment analysis."
authors = [
  { name = "mlops_imdb_team" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "dagshub>=0.6.3",
    "datasets>=3.0.0",
    "ipython",
    "joblib>=1.3",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "mlflow>=2.14,<3",
    "notebook",
    "numpy>=1.26",
    "pandas>=2.2.2",
    "pip",
    "protobuf<5",
    "pyarrow>=15",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.4",
    "tqdm",
    "typer",
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["mlops_imdb"]
include = ["pyproject.toml", "mlops_imdb/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["mlops_imdb"]
force-sort-within-sections = true

